Make generateGraphQLType internal
