+++ graphql-kotlin-schema-generator/src/test/kotlin/com/expediagroup/graphql/generator/ClassScannerTest.kt
@@ -53,46 +53,60 @@ internal class ClassScannerTest {
         abstract fun getOtherValue(): Int
     }
 
+    private annotation class SimpleAnnotation
+    private annotation class OtherSimpleAnnotation
+
+    @SimpleAnnotation
+    private class MyClassWithAnnotaiton
+
     @Test
     fun `valid subtypes`() {
-
-        val mapper = ClassScanner(listOf("com.expediagroup.graphql"))
-        val list = mapper.getSubTypesOf(MyInterface::class)
+        val classScanner = ClassScanner(listOf("com.expediagroup.graphql"))
+        val list = classScanner.getSubTypesOf(MyInterface::class)
 
         assertEquals(expected = 2, actual = list.size)
     }
 
     @Test
     fun `abstract subtypes`() {
-
-        val mapper = ClassScanner(listOf("com.expediagroup.graphql"))
-        val list = mapper.getSubTypesOf(MyAbstractClass::class)
+        val classScanner = ClassScanner(listOf("com.expediagroup.graphql"))
+        val list = classScanner.getSubTypesOf(MyAbstractClass::class)
 
         assertEquals(expected = 1, actual = list.size)
     }
 
     @Test
     fun `subtypes of non-supported packages`() {
-
-        val mapper = ClassScanner(listOf("com.example"))
-        val list = mapper.getSubTypesOf(MyInterface::class)
+        val classScanner = ClassScanner(listOf("com.example"))
+        val list = classScanner.getSubTypesOf(MyInterface::class)
 
         assertEquals(expected = 0, actual = list.size)
     }
 
     @Test
     fun `interface with no subtypes`() {
-        val mapper = ClassScanner(listOf("com.expediagroup.graphql"))
-        val list = mapper.getSubTypesOf(NoSubTypesInterface::class)
+        val classScanner = ClassScanner(listOf("com.expediagroup.graphql"))
+        val list = classScanner.getSubTypesOf(NoSubTypesInterface::class)
 
         assertEquals(expected = 0, actual = list.size)
     }
 
     @Test
     fun `abstract class with no subtypes`() {
-        val mapper = ClassScanner(listOf("com.expediagroup.graphql"))
-        val list = mapper.getSubTypesOf(NoSubTypesClass::class)
+        val classScanner = ClassScanner(listOf("com.expediagroup.graphql"))
+        val list = classScanner.getSubTypesOf(NoSubTypesClass::class)
 
         assertEquals(expected = 0, actual = list.size)
+    }
+
+    @Test
+    fun `classes with annotation returns all values`() {
+        val classScanner = ClassScanner(listOf("com.expediagroup.graphql.generator"))
+
+        val invalidClasses = classScanner.getClassesWithAnnotation(OtherSimpleAnnotation::class)
+        assertEquals(0, invalidClasses.size)
+
+        val validClasses = classScanner.getClassesWithAnnotation(SimpleAnnotation::class)
+        assertEquals(1, validClasses.size)
     }
 }
